pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.2",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.9",
    "notion-client>=2.2.1",
    "bcrypt>=5.0.0",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "google-genai>=1.62.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["google.*", "jwt", "passlib.*"]
ignore_missing_imports = true
